fix: review fixes — move mapDbCategories to module shared code
Extract DB category mapping logic from page.tsx into
modules/declaration-remuneration/shared/mapDbCategories.ts
per @maxgfr review comment (no custom logic in src/app/).
